feat(me): create common function to get assets for assets tab in loader
